Intersect PDF boxes if necessary
This reduces the PDF boxes to the MediaBox size when using the
GetPDFBoxDimensions() method.

From the PDF spec:

“The crop, bleed, trim, and art boxes should not ordinarily extend
beyond the boundaries of the media box. If the bounds of the crop,
trim, bleed or art box extends outside of the bounds of the media
box, a processor shall treat the box as its intersection with the
media box.”